do all operation as the owner

+++ src/service/handler.py
@@ -62,6 +62,9 @@ class DataHubHandler:
             password=hashlib.sha1(con_params.app_token).hexdigest(),
             repo_base=repo_base)
 
+      res = DataHubManager.has_base_privilege(user, repo_base, 'CONNECT')
+      if not (res and res['tuples'][0][0]):
+        raise Exception('Access denied. Missing required privileges.')
 
       con = Connection(
           user=user,
@@ -74,6 +77,10 @@ class DataHubHandler:
   
   def create_repo(self, con, repo_name):
     try:
+      res = DataHubManager.has_base_privilege(con.user, con.repo_base, 'CREATE')
+      if not (res and res['tuples'][0][0]):
+        raise Exception('Access denied. Missing required privileges.')
+
       manager = DataHubManager(user=con.repo_base, repo_base=con.repo_base)
       res = manager.create_repo(repo=repo_name)
       return construct_result_set(res)
@@ -82,6 +89,10 @@ class DataHubHandler:
 
   def list_repos(self, con):
     try:
+      res = DataHubManager.has_base_privilege(con.user, con.repo_base, 'CONNECT')
+      if not (res and res['tuples'][0][0]):
+        raise Exception('Access denied. Missing required privileges.')
+  
       manager = DataHubManager(user=con.repo_base, repo_base=con.repo_base)
       res = manager.list_repos()
       return construct_result_set(res)
@@ -90,6 +101,14 @@ class DataHubHandler:
 
   def delete_repo(self, con, repo_name, force_if_non_empty):
     try:
+      res = DataHubManager.has_base_privilege(con.user, con.repo_base, 'CREATE')
+      if not (res and res['tuples'][0][0]):
+        raise Exception('Access denied. Missing required privileges.')
+
+      res = DataHubManager.has_repo_privilege(con.user, con.repo_base, repo_name, 'CREATE')
+      if not (res and res['tuples'][0][0]):
+        raise Exception('Access denied. Missing required privileges.')
+
       manager = DataHubManager(user=con.repo_base, repo_base=con.repo_base)
       res = manager.delete_repo(repo=repo_name, force=force_if_non_empty)
       return construct_result_set(res)
@@ -98,6 +117,10 @@ class DataHubHandler:
 
   def list_tables(self, con, repo_name):
     try:
+      res = DataHubManager.has_repo_privilege(con.user, con.repo_base, repo_name, 'USAGE')
+      if not (res and res['tuples'][0][0]):
+        raise Exception('Access denied. Missing required privileges.')
+
       manager = DataHubManager(user=con.repo_base, repo_base=con.repo_base)
       res = manager.list_tables(repo=repo_name)
       return construct_result_set(res)
